requirements.txt
Django==1.8.1
argparse==1.2.1
wsgiref==0.1.2
MySQL-python
selenium
simplejson==2.0.9
redis==2.10.3
django-rq==0.7.0
django-rq-dashboard
django-redis==3.8.2
rq==0.4.6
rq-scheduler==0.5.0
django-bootstrap3
django-admin-bootstrapped==2.5.1
djangorestframework
gunicorn